+++ src/EFCore.Ydb/test/EntityFrameworkCore.Ydb.FunctionalTests/BulkUpdates/ComplexTypeBulkUpdatesYdbTest.cs
@@ -216,7 +216,7 @@ public class ComplexTypeBulkUpdatesYdbTest(
             """
         );
 
-    [ConditionalTheory(Skip = "Inner query contains OFFSET without LIMIT. Impossible statement in YDB")]
+    [ConditionalTheory(Skip = "YDB does not support UPDATE ... FROM (subquery) syntax yet")]
     [MemberData(nameof(IsAsyncData))]
     public override async Task Update_complex_type_to_another_database_complex_type_with_subquery(bool async)
         => await SharedTestMethods.TestIgnoringBase(
@@ -255,11 +255,6 @@ public class ComplexTypeBulkUpdatesYdbTest(
             UPDATE `Customer`
             SET `ShippingAddress_Tags` = '["new_tag1","new_tag2"]'u
             """);
-
-    [ConditionalTheory(Skip = "Inner query contains OFFSET without LIMIT. Impossible statement in YDB")]
-    [MemberData(nameof(IsAsyncData))]
-    public override Task Update_projected_complex_type_via_OrderBy_Skip(bool async)
-        => base.Update_projected_complex_type_via_OrderBy_Skip(async);
 
     public class ComplexTypeBulkUpdatesYdbFixture : ComplexTypeBulkUpdatesRelationalFixtureBase
     {